% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changetitle.R
\name{changetitle}
\alias{changetitle}
\title{changetitle}
\usage{
changetitle(titl)
}
\arguments{
\item{titl}{vector contains titles that need to be reshaped}
}
\description{
Reshape the title for plots
}
\details{
This funciton reshapes a vector of characters inside. This vector often serves as a list of titles
for different plots. The aim for this function is to narrow down the length of each element in this vector
so that the name will fit into plots. This function will put a line break(i.e. "\\n") each time after counting
25 characters.
}
\examples{
head(dataColnames)
newtitles<-changetitle(dataColnames)
head(newtitles)

}
\author{
SOCR team <http://socr.umich.edu/people/>
}
