% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleardata.R
\name{cleardata}
\alias{cleardata}
\title{cleardata}
\usage{
cleardata(mat)
}
\arguments{
\item{mat}{A matrix to be cleaned.}
}
\value{
A matrix whose NA's are all filled with column mean of the rest items.
}
\description{
This function imputes missing values for a dataset
}
\details{
This function use the mean value of each column to impute each missing value (i.e. NA value)
in that column. Thus provides a quick and easy method to resolve value missing problem in a dataset.
}
\examples{
#Need package "VIM" to make this example work
require(VIM)

# See the missing value in dataset "sleep"
head(VIM:sleep)

# Deal with missing values
cleardata(VIM:sleep)

}
\author{
SOCR team <http://socr.umich.edu/people/>
}
