% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completeHomologousX_features.R
\name{completeHomologousX_features}
\alias{completeHomologousX_features}
\title{completeHomologousX_features}
\usage{
completeHomologousX_features(list_of_dfs)
}
\arguments{
\item{list_of_dfs}{a list of datasets of different countries}
}
\description{
Ensure the XReg predictors are homologous
}
\details{
This function ensures the XReg predictors for ALL 31 EU countries are homologous
\strong{Notice:} This function mainly serves as a functional part for other functions. So we
seldomly apply this function by itself.
}
\author{
SOCR team <http://socr.umich.edu/people/>
}
