% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countryFeature2Index.R
\name{countryFeature2Index}
\alias{countryFeature2Index}
\title{countryFeature2Index}
\usage{
countryFeature2Index(countryIndx = 1, featureIndx = 1)
}
\arguments{
\item{countryIndx}{indicate the country you wish to transfer}

\item{featureIndx}{indicate the feature you wish to transfer}
}
\description{
Convert country feature to index
}
\details{
This function transfer country and feature code to 1D index
An opposite procedure of function "index2CountryFeature"
\strong{Notice:} This function mainly serves as a functional part for other functions. So we seldomly
apply this function by itself.
}
\author{
SOCR team <http://socr.umich.edu/people/>
}
