% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fftshift.R
\name{fftshift}
\alias{fftshift}
\title{fftshift}
\usage{
fftshift(img_ff, dim = -1)
}
\arguments{
\item{img_ff}{A Fourier transform of a 1D signal, 2D image, or 3D volume.}

\item{dim}{Number of dimensions (-1, 1, 2, 3).}
}
\value{
A properly shifted FT of the array.
}
\description{
FFT SHIFT
}
\details{
This function is useful for visualizing the Fourier transform with the zero-frequency
component in the middle of the spectrum.
}
\examples{
fftshift(ft_circle)

}
\author{
SOCR team <http://socr.umich.edu/people/>
}
