% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fftshift1D.R
\name{fftshift1D}
\alias{fftshift1D}
\title{A special case of imlementaiton of `fftshift` for 1D arrays}
\usage{
fftshift1D(img_ff)
}
\arguments{
\item{img_ff}{A Fourier transform of a 1D signal.}
}
\value{
A properly shifted FT of the 1D array.
}
\description{
This function is useful for visualizing the 1D Fourier transform with the zero-frequency
component in the middle of the spectrum.
}
\examples{
plot(fftshift1D(log(Re(X2)+2)), main = "log(fftshift1D(Re(FFT(timeseries))))")

}
\author{
SOCR team <http://socr.umich.edu/people/>
}
