% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_simulate_func.R
\name{fmri_simulate_func}
\alias{fmri_simulate_func}
\title{real-valued fMRI data simulation}
\usage{
fmri_simulate_func(
  dim_data,
  mask = NULL,
  ons = c(1, 21, 41, 61, 81, 101, 121, 141),
  dur = c(10, 10, 10, 10, 10, 10, 10, 10)
)
}
\arguments{
\item{dim_data}{a vector of length three to identify the dimension of fMRI data wanted to simulate}

\item{mask}{a parameter to have the fMRI data with the brain-shaped mask provided in the package, or the 3d sphere mask 
generated by the function itself.}

\item{ons}{a vector of the start time points of the time period when the fMRI data receives stimulation}

\item{dur}{a vector of the time period when the fMRI data receives stimulation}
}
\value{
an array with the specified dimension

a list of four elements
\itemize{
  \item fmri_data - the fMRI data generated by the function as specialized values.
  \item mask - mask of the fMRI data.
  \item ons - a vector of the start time points of the time period when the fMRI data receives stimulation.
  \item dur - a vector of the time period when the fMRI data receives stimulation. Notice that the length of ons 
  is equal to the length of dur, and all the time period when the data does not receive the simulations have the same 
  duration as its former 'on' time period.
}
}
\description{
an fMRI data simulation function, used to simply generate a 3 dimension area with activated parts inside.
}
\details{
The function \code{fmri_simulate_func} is used to simulate fMRI data with specified dimension and total time points. 
The fMRI data can own a brain shape by using the mask data provided in our package, if the dimension fits the same as our data (c(64, 64. 40)).
Otherwise, the function will generate a 3d sphere data with multiple activated part inside. The activated parts can be detected based on the p values.
}
\examples{
# sample 3D data of mask provided by the package
dim(mask)

# the input dimension is the dimension we want for our simulated fMRI data
fmri_generate = fmri_simulate_func(dim_data = c(64, 64, 40), mask = mask, 
                                   ons = c(1, 21, 41, 61, 81, 101, 121, 141), 
                                   dur = c(10, 10, 10, 10, 10, 10, 10, 10))

}
\author{
SOCR team <\url{http://socr.umich.edu/people/}>
}
