% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GaussSmoothKernel.R
\name{GaussSmoothKernel}
\alias{GaussSmoothKernel}
\title{GaussSmoothKernel}
\usage{
GaussSmoothKernel(voxdim = c(1, 1, 1), ksize = 5, sigma = diag(3, 3))
}
\arguments{
\item{voxdim}{Dimensions of each voxel.}

\item{ksize}{Dimensions of the discrete kernel size.}

\item{sigma}{The covariance matrix of the Gaussian kernel.}
}
\value{
a 3 dimensional array with size = (ksize, ksize, ksize)
}
\description{
An internal function named GaussSmoothKernel. Original from AnalyzeFMRI package
}
