% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index2CountryFeature.R
\name{index2CountryFeature}
\alias{index2CountryFeature}
\title{index2CountryFeature}
\usage{
index2CountryFeature(indx = 1)
}
\arguments{
\item{indx}{Indicate the feature you wish to see after transfering to 1D index}
}
\value{
return the original country and feature code of the feature you wish to show
}
\description{
Convert index to country name and feature code
}
\details{
This function maps to convert between 1D indices
\strong{Notice:} This function mainly serves as a functional part for other functions. So we seldomly
apply this function by itself.
}
\author{
SOCR team <http://socr.umich.edu/people/>
}
