% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LT.R
\name{LT}
\alias{LT}
\title{numerical method to compute Laplace Transform}
\usage{
LT(FUNCT, z)
}
\arguments{
\item{FUNCT}{a function object f(t) conducting Laplace Transform}

\item{z}{a complex domain value used to evaluate the F(z)=LT(f)(z)}
}
\value{
a complex value computed from Laplace Transform
}
\description{
a function that numerically computes the Laplace Transform
}
\examples{
f = function(t) { t }; z= 1+1i; 
LT(f, z); 
# compare with the result from analytic form of Laplace Transform of f(t) = t
# analytic form is below
F = function (z) { 1/z^2 }; F(z)
# the two results are the same
}
\author{
SOCR team <\url{http://socr.umich.edu/people/}>
}
