% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_dict.R
\docType{data}
\name{mask_dict}
\alias{mask_dict}
\title{mask_dict}
\format{
a data.frame containing the label index corresponding to its label name
}
\usage{
mask_dict
}
\description{
a data.frame containing the label index corresponding to its label name
}
\keyword{data.frame}
