% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_label.R
\docType{data}
\name{mask_label}
\alias{mask_label}
\title{mask_label}
\format{
a 3D array containing brain mask data
}
\usage{
mask_label
}
\description{
a 64*64*40 3D array representing brain mask with labels
}
\keyword{array}
