% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedDistFunc.R
\name{mixedDistFunc}
\alias{mixedDistFunc}
\title{mixedDistFunc}
\usage{
mixedDistFunc(n, weight, mu1, mu2, sig1, sig2)
}
\arguments{
\item{n}{A integer, specifies number of random numbers generated by each normal distribution}

\item{weight}{Weight of the first normal distribution}

\item{mu1}{Mean of the first normal distribution}

\item{mu2}{Mean of the secondnormal distribution}

\item{sig1}{Standard deviation of the first normal distribution}

\item{sig2}{Standard deviation of the second normal distribution}
}
\value{
A number or numbers that is a combination of two normal distributions
}
\description{
This function generates a new distribution that is mixed model of two normal distributions
}
\details{
This function takes in two pairs of mean and standard deviation to be used in two gaussian
distributions. In addition, a weight parameter is also needed to adjust the proportion of both gaussian.
}
\examples{
n=10000;mu1 <- 0; mu2 <- 5;sig1 <- 1; sig2 <- 3;weight <- 0.7
xA <- mixedDistFunc(n=n, weight, mu1, mu2, sig1, sig2)
hist(xA, freq = F)
}
\author{
SOCR team <http://socr.umich.edu/people/>
}
