% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase1_pval.R
\docType{data}
\name{phase1_pval}
\alias{phase1_pval}
\title{phase1_pval}
\format{
a 3D array containing p values
}
\usage{
phase1_pval
}
\description{
a 64*64*40 3D array containing a sample p values for the first phase of three-phase ROI analysis by function fmri_ROI_phase1
}
\keyword{array}
