% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase2_pval.R
\docType{data}
\name{phase2_pval}
\alias{phase2_pval}
\title{phase2_pval}
\format{
a 3D array containing p values
}
\usage{
phase2_pval
}
\description{
a 64*64*40 3D array containing a sample p values for the second phase of three-phase ROI analysis by function fmri_ROI_phase2
}
\keyword{array}
