% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase3_pval.R
\docType{data}
\name{phase3_pval}
\alias{phase3_pval}
\title{phase3_pval}
\format{
a 3D array containing p values
}
\usage{
phase3_pval
}
\description{
a 64*64*40 3D array containing a sample p values for the third phase of three-phase ROI analysis, generated by the post-hoc process for phase2_pval
}
\keyword{array}
