% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_ARIMA.R
\name{preprocess_ARIMA}
\alias{preprocess_ARIMA}
\title{preprocess_ARIMA}
\usage{
preprocess_ARIMA(country = "Belgium", start = 2000, end = 2017,
  frequency = 20,
  feature = "Unemployment , Females, From 15-64 years, Total")
}
\arguments{
\item{country}{Give a country name that is shown on the original dataset}

\item{start}{Select the start year to create the ARIMA model}

\item{end}{Select the end year to create the ARIMA model}

\item{frequency}{The number of observations per unit of time. The same in function "ts"}

\item{feature}{Choose one feature to create the ARIMA model on}
}
\value{
a list contains two elements:
\item{X}{A data frame contains the full rank part of all features}
\item{Y}{A vector contains response variable}
}
\description{
Preprocess dataset for ARIMA model
}
\details{
This function preprocesses dataset of given countries to ensure full rank
Then, extend the Fit-ARIMA method to ensure testing-training
modeling/assessment for 2 countries works
}
\examples{
preprocess_Belgium <- preprocess_ARIMA( country = 'Belgium', 
start=2000, end=2017, frequency=20, 
feature="Gross domestic product at market prices")

# Just show the response variable part
head(preprocess_Belgium$Y)
}
\author{
SOCR team <http://socr.umich.edu/people/>
}
