% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval1.R
\docType{data}
\name{pval1}
\alias{pval1}
\title{pval1}
\format{
a 3D array containing p values
}
\usage{
pval1
}
\description{
a 64*64*40 3D array containing p values, generated by the real fMRI data with the hemodynamic response function method
}
\keyword{array}
