% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval2.R
\docType{data}
\name{pval2}
\alias{pval2}
\title{pval2}
\format{
a 3D array containing p values
}
\usage{
pval2
}
\description{
a 64*64*40 3D array containing p values, generated by the simulated fMRI data with the t-test method
}
\keyword{array}
