% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmv_miss_ftr.R
\name{rmv_miss_ftr}
\alias{rmv_miss_ftr}
\title{rmv_miss_ftr}
\usage{
rmv_miss_ftr(countryName = "Bulgaria")
}
\arguments{
\item{countryName}{give a country name that is shown on the original dataset}
}
\description{
Removes features with all missing values
}
\details{
This function deletes features that are missing for all time points
associated with a certain country.
}
\examples{
require(dplyr)

#take "Sweden" as an example
head(rmv_miss_ftr("Sweden"))[,1:10]
}
\author{
SOCR team <http://socr.umich.edu/people/>
}
