% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\docType{data}
\name{sample}
\alias{sample}
\title{sample}
\format{
a data.frame containing some pre-calculated data for generating vignettes
}
\usage{
sample
}
\description{
a data.frame containing some pre-calculated data for generating vignettes
}
\keyword{data.frame}
