% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_save.R
\docType{data}
\name{sample_save}
\alias{sample_save}
\title{sample_save}
\format{
a list containing some pre-calculated data for generating vignettes
}
\usage{
sample_save
}
\description{
a list containing some pre-calculated data for generating vignettes
}
\keyword{list}
